Refactor data loading and sidebar layout
