+++ src/utils/format.ts
@@ -1,0 +1,5 @@
+const numberFormatter = new Intl.NumberFormat("en-US");
+
+export const formatNumber = (value: number): string => numberFormatter.format(value);
+
+export const normalizeKey = (value: string): string => value.trim().toLowerCase();
